requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
httpx==0.27.2
cryptography==43.0.3
sqlalchemy==2.0.36
alembic==1.13.2
psycopg2-binary==2.9.10
redis==5.0.8

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.3.0
python-multipart==0.0.9
email-validator==2.1.0

# Google APIs
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.148.0

# Optional LLM
openai==1.51.2